+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-8257 Added closing modal on transition
 * VFS-11006 Changed EGI auth icon
 * VFS-10563 Added record info with icon component for providers
 * VFS-10895 Added expand options to automation data spec of file